Reorganize components directory and switch to using absolute imports
