+++ network-topology/src/App.js
@@ -1,12 +1,12 @@
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import './App.css';
-import Grid from "./components/Grid";
-import GridPage from "./components/GridPage";
-import TransformerPage from "./components/TransformerPage";
-import NetworkTopology from "./components//Network/NetworkTopology";
-import UserConfiguration from "./components/HouseConfiguration";
-import LoadBuilder from "./components/LoadProfile/LoadBuilder";
-import GenerationPage1 from "./components/LoadProfile/GenerationPage1";
+import Grid from "components/Grid/Grid";
+import GridPage from "components/Grid/GridPage";
+import TransformerPage from "components/Transformer/TransformerPage";
+import NetworkTopology from "components/Network/NetworkTopology";
+import UserConfiguration from "components/House/HouseConfiguration";
+import LoadBuilder from "components/LoadProfile/LoadBuilder";
+import GenerationPage1 from "components/LoadProfile/GenerationPage1";
 
 function App() {
   return (
