+++ network-topology/src/components/Common/DeleteConfirm.jsx
@@ -1,13 +1,12 @@
 import React from "react";
-import Breadcrumb from "./Breadcrumb";
-import "./Breadcrumb.css"
+import Breadcrumb from "components/Breadcrumb/Breadcrumb";
 
 const Delete = ({ onConfirm, onClose, entity, entityId, entityName, entityType }) => {
   return (
     <div className="fixed inset-0 flex items-center justify-center bg-black bg-opacity-50 z-50">
       <div className="relative flex flex-col justify-center text-xl font-medium text-orange-600 w-[800px] max-w-[800px] bg-white rounded-3xl border border-sky-400 border-solid shadow-sm z-10">
         <div className="absolute top-5 z-1 text-[14px] text-black font-light">
-          {entity && entity.new!==true && <Breadcrumb nodeId={entityId} onEditNode={() => {}}/>}
+          {entity && entity.new !== true && <Breadcrumb nodeId={entityId} onEditNode={() => { }} />}
         </div>
         <div className="flex flex-col items-center  pb-14 pt-2  max-md:pr-5 max-md:max-w-full">
           <button className="cursor-pointer absolute top-3 right-5 p-2 text-4xl font-thin text-black" onClick={onClose}>X</button>
@@ -17,9 +16,9 @@ const Delete = ({ onConfirm, onClose, entity, entityId, entityName, entityType }
           </div>
           <div className="mt-8 text-center max-md:mt-10"> Please Confirm</div>
           <button
-           className="px-20 py-1 mt-16 text-[20px] h-[55px] w-[300px] font-bold text-center text-navColor bg-amber-400 rounded-2xl max-md:px-5 max-md:mt-10" onClick={onConfirm}
+            className="px-20 py-1 mt-16 text-[20px] h-[55px] w-[300px] font-bold text-center text-navColor bg-amber-400 rounded-2xl max-md:px-5 max-md:mt-10" onClick={onConfirm}
           >
-           YES
+            YES
           </button>
         </div>
       </div>
